EditTags navigation steps cleanup/move
