+++ cfme/cloud/flavor.py
@@ -5,7 +5,7 @@ from widgetastic.exceptions import NoSuchElementException
 from widgetastic_patternfly import Dropdown, Button, View
 
 from cfme.base.ui import BaseLoggedInPage
-from cfme.common import TagPageView, WidgetasticTaggable
+from cfme.common import WidgetasticTaggable
 from cfme.exceptions import FlavorNotFound
 from cfme.web_ui import match_location
 from cfme.utils.appliance import Navigatable
@@ -124,12 +124,3 @@ class FlavorDetails(CFMENavigateStep):
             raise FlavorNotFound('Could not locate flavor "{}" on provider {}'
                                  .format(self.obj.name, self.obj.provider.name))
         row.click()
-
-
-@navigator.register(Flavor, 'EditTagsFromDetails')
-class FlavorEditTags(CFMENavigateStep):
-    VIEW = TagPageView
-    prerequisite = NavigateToSibling('Details')
-
-    def step(self, *args, **kwargs):
-        self.prerequisite_view.toolbar.policy.item_select('Edit Tags')
