+++ cfme/networks/subnet.py
@@ -1,6 +1,6 @@
-from navmazing import NavigateToSibling, NavigateToAttribute
+from navmazing import NavigateToAttribute
 
-from cfme.common import WidgetasticTaggable, TagPageView
+from cfme.common import WidgetasticTaggable
 from cfme.exceptions import ItemNotFound
 from cfme.networks.views import SubnetDetailsView, SubnetView
 from cfme.utils import providers, version
@@ -88,12 +88,3 @@ class OpenCloudNetworks(CFMENavigateStep):
 
     def step(self):
         self.prerequisite_view.entities.get_entity(by_name=self.obj.name).click()
-
-
-@navigator.register(Subnet, 'EditTagsFromDetails')
-class EditTags(CFMENavigateStep):
-    prerequisite = NavigateToSibling('Details')
-    VIEW = TagPageView
-
-    def step(self):
-        self.prerequisite_view.toolbar.policy.item_select('Edit Tags')
